move all weibull pieces into one
